cleaning up tests, renaming to scriptify, adding extension functionality
